Revert "rw spin lock"
